[refactor] controller 설정 디렉토리
