+++ src/main/java/com/jinddung2/givemeticon/common/config/controller/response/ApiResponse.java
@@ -1,19 +1,19 @@
-package com.jinddung2.givemeticon.common.response;
+package com.jinddung2.givemeticon.common.config.controller.response;
 
 public record ApiResponse<T>(
         String message,
         T data
 ) {
-    public static <T> ApiResponse<T> success() {
-        return new ApiResponse<>("SUCCESS", null);
+    public static <T> ApiResponse<?> success() {
+        return new ApiResponse<>("SUCCESS", "no data");
     }
 
     public static <T> ApiResponse<T> success(T result) {
         return new ApiResponse<>("SUCCESS", result);
     }
 
-    public static <T> ApiResponse<T> fail() {
-        return new ApiResponse<>("FAIL", null);
+    public static <T> ApiResponse<?> fail() {
+        return new ApiResponse<>("FAIL", "no data");
     }
 
     public static <T> ApiResponse<T> fail(T result) {
