+++ src/main/java/com/jinddung2/givemeticon/domain/category/advice/CategoryExceptionAdvice.java
@@ -1,7 +1,7 @@
 package com.jinddung2.givemeticon.domain.category.advice;
 
-import com.jinddung2.givemeticon.common.response.ApiResponse;
-import com.jinddung2.givemeticon.common.response.ErrorResult;
+import com.jinddung2.givemeticon.common.config.controller.response.ApiResponse;
+import com.jinddung2.givemeticon.common.config.controller.response.ErrorResult;
 import com.jinddung2.givemeticon.domain.category.exception.CategoryException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
