+++ src/main/java/com/jinddung2/givemeticon/domain/sale/advice/SaleExceptionAdvice.java
@@ -1,7 +1,7 @@
 package com.jinddung2.givemeticon.domain.sale.advice;
 
-import com.jinddung2.givemeticon.common.response.ApiResponse;
-import com.jinddung2.givemeticon.common.response.ErrorResult;
+import com.jinddung2.givemeticon.common.config.controller.response.ApiResponse;
+import com.jinddung2.givemeticon.common.config.controller.response.ErrorResult;
 import com.jinddung2.givemeticon.domain.sale.exception.SaleException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
